backend/requirements.txt
Django==6.0
djangorestframework
playwright
pandas
django-cors-headers
nest_asyncio